had issues with commits
i had issues where the last version got corrupted and i had to use the last backed up version of my project. i tried to commit but it kept saying that my code is behind the branch im trying to commit to. so im forced to upload zip files. i have tried to push again but it isnt working